routing and don't remove people from finished games

+++ src/app/components/game-join-button/game-join-button.component.ts
@@ -2,7 +2,7 @@ import {Component, Input} from '@angular/core';
 import {Router} from '@angular/router';
 import {AuthService} from 'src/app/services/auth.service';
 import {RoomService} from 'src/app/services/room.service';
-import {Game, Room} from 'types';
+import {Game, Room, RoomStatus} from 'types';
 
 import {GameService} from '../../services/game.service';
 
@@ -51,14 +51,19 @@ export class GameJoinButtonComponent {
   }
 
   leave() {
+    // remove them from the room
     this.roomService.removePlayerFromRoom(
         this.room.id,
         this.authService.currentUserId,
     );
-    this.gameService.removePlayerFromGame(
-        this.game.id,
-        this.authService.currentUserId,
-    );
-    this.router.navigate(['home']);
+    // remove them from the game, so long as it hasn't already completed
+    if (!this.game.completedAt) {
+      this.gameService.removePlayerFromGame(
+          this.game.id,
+          this.authService.currentUserId,
+      );
+    }
+    // go home
+    this.router.navigate(['']);
   }
 }
